Fix merging devDependencies (#43)
I noticed that the devDependency section of package.json was not being merged into the thirdPartyModules object correctly.

I added a unit test for the module, as well as changed the test for transform.js to use a mocked package.json for better test isolation, and updated the test fixtures to exercise the code with that mocked data. Additionally, I have updated the README to specify that dev dependencies are included alongside the normal dependencies.